implement motley/vendor structure
