--- api/Ourchitecture.Api.Grasshopper/Class1.cs
@@ -1,12 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace Ourchitecture.Api.Grasshopper
-{
-    public class Class1
-    {
-    }
-}
