+++ api/Ourchitecture.Api.Grasshopper/packages.config
@@ -1,5 +1,5 @@
 ﻿<?xml version="1.0" encoding="utf-8"?>
 <packages>
-  <package id="Grasshopper" version="6.0.18016.23451" targetFramework="net461" />
-  <package id="RhinoCommon" version="6.0.18016.23451" targetFramework="net461" />
+  <package id="Grasshopper" version="6.12.19029.6381" targetFramework="net461" />
+  <package id="RhinoCommon" version="6.12.19029.6381" targetFramework="net461" />
 </packages>
