+++ api/Ourchitecture.Api/Protocols/Motley/MotleyModule.cs
@@ -5,20 +5,21 @@ using System.Web;
 using Nancy;
 using Rhino.Geometry;
 using Newtonsoft.Json;
+using Ourchitecture.Api.Protocols.Motley;
 
-namespace Ourchitecture.Api.Protocols
+namespace Ourchitecture.Api.Protocols.Motley
 {
     public class MotleyModule : NancyModule
     {
-        private Dictionary<string, Action> MotleyRoutes { get; } = new Dictionary<string, Action>()
+        private Dictionary<string, Func<string, string>> MotleyRoutes { get; } = new Dictionary<string, Func<string, string>>()
         {
-
+            { "vendor", RunVendor },
         };
 
         public MotleyModule()
         {
             Get["/motley"] = _ => MotleyManifest();
-            Post["/motley/{dialect}"] = parameters => MotleyRoutes[parameters.dialect];
+            Post["/motley/{dialect}"] = parameters => MotleyRoutes[parameters.dialect](Request.Body.ToString());
         }
 
         private string MotleyManifest()
@@ -27,6 +28,11 @@ namespace Ourchitecture.Api.Protocols
             return JsonConvert.SerializeObject(crv);
             
             //return JsonConvert.SerializeObject(MotleyRoutes.Keys.Select(x => x));
+        }
+
+        private static string RunVendor(string request)
+        {
+            return VendorSchema.Solve(new VendorRequest(request));
         }
     }
 }
